Code Improvement

+++ ItemManagementSystem.Application/Implementation/ItemTypeService.cs
@@ -5,6 +5,7 @@ using ItemManagementSystem.Application.Interface;
 using ItemManagementSystem.Domain.Constants;
 using ItemManagementSystem.Domain.DataModels;
 using ItemManagementSystem.Domain.Dto;
+using ItemManagementSystem.Domain.Dto.Request;
 using ItemManagementSystem.Domain.Exception;
 using ItemManagementSystem.Infrastructure.Interface;
 
@@ -40,14 +41,23 @@ namespace ItemManagementSystem.Application.Implementation
             return userId;
         }
 
-        public async Task<ItemTypeDto> CreateAsync(ItemTypeDto dto)
+        public async Task<ItemTypeCreateRequest> CreateAsync(ItemTypeCreateRequest dto, int userId)
         {
-            dto.Id=0;
-            dto.modifiedBy=null;
-            
+            if (dto == null)
+                throw new NullObjectException(AppMessages.NullItemTypeRequest);
+
+            var exists = (await _repo.FindAsync(
+             it => it.Name.ToLower() == dto.Name.ToLower()
+         )).Any();
+
+            if (exists)
+                throw new AlreadyExistsException(AppMessages.ItemTypeAlreadyExists);
+
             var entity = _mapper.Map<ItemType>(dto);
+            entity.CreatedBy = userId;
+
             var created = await _repo.AddAsync(entity);
-            return _mapper.Map<ItemTypeDto>(created);
+            return _mapper.Map<ItemTypeCreateRequest>(created);
         }
 
         public async Task<ItemTypeDto?> GetByIdAsync(int id)
@@ -55,27 +65,45 @@ namespace ItemManagementSystem.Application.Implementation
             var entity = await _repo.GetByIdAsync(id);
             if (entity == null)
                 throw new NullObjectException(AppMessages.ItemTypeNotFound);
-            return  _mapper.Map<ItemTypeDto>(entity);
+
+            if(entity.IsDeleted)
+                throw new NullObjectException(AppMessages.ItemTypeNotFound);    
+            return _mapper.Map<ItemTypeDto>(entity);
         }
 
         public async Task<IEnumerable<ItemTypeDto>> GetAllAsync()
         {
             var entities = await _repo.GetAllAsync();
-            return _mapper.Map<IEnumerable<ItemTypeDto>>(entities);
+
+            var filtered = entities
+                .Where(e => !e.IsDeleted)  
+                .ToList();
+
+            return _mapper.Map<IEnumerable<ItemTypeDto>>(filtered);
         }
 
-        public async Task<ItemTypeDto> UpdateAsync(int id, ItemTypeDto dto)
+
+        public async Task<ItemTypeCreateRequest> updateAsync(int id, ItemTypeCreateRequest dto, int userId)
         {
-            dto.Id=id;
+
             var entity = await _repo.GetByIdAsync(id);
-            if (entity == null)
+            if (entity == null || entity.IsDeleted)
                 throw new NullObjectException(AppMessages.ItemTypeNotFound);
 
-            dto.createdBy= entity.CreatedBy;    
+            //if same name exist
+            var exists = (await _repo.FindAsync(
+                it => it.Name.ToLower() == dto.Name.ToLower() && it.Id != id
+            )).Any();
+
+            if (exists)
+                throw new AlreadyExistsException(AppMessages.ItemTypeAlreadyExists);
+
 
             _mapper.Map(dto, entity);
+            entity.UpdatedAt = DateTime.UtcNow;
+            entity.ModifiedBy = userId;
             await _repo.UpdateAsync(entity);
-            return _mapper.Map<ItemTypeDto>(entity);
+            return _mapper.Map<ItemTypeCreateRequest>(entity);
         }
 
         public async Task DeleteAsync(int id)
@@ -83,6 +111,12 @@ namespace ItemManagementSystem.Application.Implementation
             var entity = await _repo.GetByIdAsync(id);
             if (entity == null)
                 throw new NullObjectException(AppMessages.ItemTypeNotFound);
+
+            var hasAssociatedModels = (await _repo.FindAsync(
+                it => it.ItemModels.Any(im => im.IsDeleted == false)
+            )).Any();
+            if (hasAssociatedModels)
+                throw new CustomException(AppMessages.ItemTypeHasAssociatedModels);
 
             await _repo.DeleteAsync(entity);
         }
